Aqua-Flare-Robot- 405 Found/backend/requirements.txt
attrs==25.1.0
bcrypt==4.2.1
blinker==1.9.0
click==8.1.8
dnspython==2.7.0
flasgger==0.9.7.1
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
mistune==3.1.1
packaging==24.2
pickle-mixin==1.0.2
pymongo==4.11
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.22.3
six==1.17.0
Werkzeug==3.1.3
